requirements.txt
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Django==4.0.1
djangorestframework==3.13.1
h11==0.12.0
idna==3.3
importlib-metadata==4.10.0
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
pyee==8.2.2
pyOpenSSL==21.0.0
pyppeteer==1.0.2
pytz==2021.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==1.2.3
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
websockets==10.1
wsproto==1.0.0
zipp==3.7.0
